use of atomicAdd operation; working for medium pangemes (e.g., MHCX4)
small pangenomes are not working: Their coordinates seem to be corrupted
because so many different cores update their coordinates in parallel.
use of floats, because atomicAdd does not want to work with doubles.

+++ src/cuda/layout.h
@@ -22,7 +22,7 @@ namespace cuda {
 
 struct node_t {
     int32_t seq_length;
-    double coords[4];
+    float coords[4];
 };
 struct node_data_t {
     uint32_t node_count;
